Move migrations runner out of app.py
